# Switch from m3db/stackmurmur3 to a customized implementation of murmur3