--- a/glide.lock
+++ b/glide.lock
@@ -1,19 +1,20 @@
-hash: 92e25ff1ff7d4e7802b22bbef0dda7a873d90ef3823bef0e0ba859364cd63a88
-updated: 2020-08-27T14:37:28.119756-04:00
+hash: 2b9862f9963ea7bd113a42f2787b4921cc3646d9bf49bb58fd3c4c49c91714f0
+updated: 2020-08-27T21:13:59.818543-04:00
 imports:
 - name: github.com/m3db/bitset
   version: 07973db6b78acb62ac207d0538055e874b49d90d
-- name: github.com/m3db/stackmurmur3
-  version: 744c0229c12ed0e4f8cb9d081a2692b3300bf705
-- name: github.com/spaolacci/murmur3
-  version: 0d12bf811670bf6a1a63828dfbd003eded177fce
-- name: github.com/twmb/murmur3
-  version: a57a9e6e8d2a5858df4163dd85863a6a9d9c5474
 testImports:
 - name: github.com/davecgh/go-spew
   version: 5215b55f46b2b919f50a1df0eaa5886afe4e3b3d
   subpackages:
   - spew
+- name: github.com/leanovate/gopter
+  version: 8de0d746533ed4ca265750b9bcf3ba01839df390
+  subpackages:
+  - gen
+  - prop
+- name: github.com/m3db/stackmurmur3
+  version: 744c0229c12ed0e4f8cb9d081a2692b3300bf705
 - name: github.com/pmezard/go-difflib
   version: d8ed2627bdf02c080bf22230dbb337003b7aba2d
   subpackages:
@@ -23,3 +24,5 @@ testImports:
   subpackages:
   - assert
   - require
+- name: github.com/twmb/murmur3
+  version: a57a9e6e8d2a5858df4163dd85863a6a9d9c5474