--- a/murmur.go
+++ b/murmur.go
@@ -24,8 +24,9 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 		h1, h2 uint64
 		res    [4]uint64
 	)
+	// Calculate 1st set of hashes - twmb/murmur3 Sum128() of payload
 	clen := len(data)
-	for len(data) >= 16 {
+	for len(data) >= 16 { // bmix() step
 		// yes, this is faster than using binary.LittleEndian.Uint64
 		k1 := uint64(data[0]) | uint64(data[1])<<8 | uint64(data[2])<<16 | uint64(data[3])<<24 | uint64(data[4])<<32 | uint64(data[5])<<40 | uint64(data[6])<<48 | uint64(data[7])<<56
 		k2 := uint64(data[8]) | uint64(data[9])<<8 | uint64(data[10])<<16 | uint64(data[11])<<24 | uint64(data[12])<<32 | uint64(data[13])<<40 | uint64(data[14])<<48 | uint64(data[15])<<56
@@ -51,7 +52,11 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 		h2 = h2*5 + 0x38495ab5
 	}
 
-	var _h1, _h2 = h1, h2 // save state after mixing, before final processing
+	// The following line diverges from twmb/murmur3.Sum128() in a cruicial way:
+	// - for the 2nd hash set for Bloom filter, we need to save seed state to recalculate the hash
+	//   were the *data* was exactly the same, just with the last *byte* being equal to `entropy`
+	var _h1, _h2 = h1, h2
+	// end of change from twmb/murmur3
 	var k1, k2 uint64
 	switch len(data) {
 	case 15:
@@ -122,16 +127,27 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 
 	h1 += h2
 	h2 += h1
+	// End of original twmb/murmur3.Sum128
+	// Save this hash, as it's equivalent of h1, h2 = hash.Write(data).Sum128()
+	res[0], res[1] = h1, h2
 
-	res[0], res[1] = h1, h2 // done with 1st hash
-	// the last 128 bytes is hash for the same data with 0x01 appended for entropy
-	// h3, h4 = hash.Write(data).Write([]byte{1}).Sum128()
+	// Now, reset state back to before the tail calculation.
+	// The state is such that all but up to final 15 bytes have been processed, as murmur3 works
+	// on 128 bit / 16 byte chunks for all but the final mixing.
+	// last byte of the array, and apply the Sum128() logic accordingly.
 	k1, k2 = 0, 0
-	h1, h2 = _h1, _h2      // restore state
-	clen++                 // add entropy byte to total hash data length
-	switch len(data) + 1 { // process entropy byte *only*, no fallthrough - unless it's now a full block
-	case 16: // == entire block
+	h1, h2 = _h1, _h2 // restore state to before final mixing
+	clen++            // total hash data length now includes entropy
+	// Fist, process entropy byte *only*. No fallthrough.
+	// However, if we are really lucky and remainder was 15 bytes, with entropy byte it is now a full block.
+	// In that case, just apply bmix() again, with the modified payload and directly go to hash finalization.
+	rlen := len(data)
+
+	// Check if we're lucky, and have to just perform bmix() of the final chunk
+	if rlen == 15 {
 		k1 = uint64(data[0]) | uint64(data[1])<<8 | uint64(data[2])<<16 | uint64(data[3])<<24 | uint64(data[4])<<32 | uint64(data[5])<<40 | uint64(data[6])<<48 | uint64(data[7])<<56
+		// Note: this is different from original bmix() - we don't mutate `data`,
+		// so we reuse the entire 15 byte slice of the original remainder and treat the last byte as equal to `entropy`
 		k2 = uint64(data[8]) | uint64(data[9])<<8 | uint64(data[10])<<16 | uint64(data[11])<<24 | uint64(data[12])<<32 | uint64(data[13])<<40 | uint64(data[14])<<48 | entropy<<56
 
 		k1 *= c1_128
@@ -151,6 +167,13 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 		h2 = bits.RotateLeft64(h2, 31)
 		h2 += h1
 		h2 = h2*5 + 0x38495ab5
+		// No further processing needed, skip the next state
+		goto Finalize
+	}
+
+	// Process entropy byte - note, **no fallthrough**, because
+	// how we process it depends on how long the payload is
+	switch rlen + 1 {
 	case 15:
 		k2 ^= entropy << 48
 	case 14:
@@ -184,15 +207,17 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 	case 2:
 		k1 ^= entropy << 8
 	case 1:
+		// Case where entropy byte is the actual tail (payload length is a multiple of 16 bytes)
 		k1 ^= entropy << 0
 		k1 *= c1_128
 		k1 = bits.RotateLeft64(k1, 31)
 		k1 *= c2_128
 		h1 ^= k1
+		// Nothing else to do but finalize the hash
+		goto Finalize
 	}
 
-	switch len(data) {
-	case 15: // nothing, entire block processed
+	switch rlen {
 	case 14:
 		k2 ^= uint64(data[13]) << 40
 		fallthrough
@@ -247,6 +272,7 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 		h1 ^= k1
 	}
 
+Finalize:
 	h1 ^= uint64(clen)
 	h2 ^= uint64(clen)
 
@@ -258,6 +284,8 @@ func sum128WithEntropy(data []byte) [4]uint64 {
 
 	h1 += h2
 	h2 += h1
+
+	// Save 2nd hash, twmb/murmur3.Sum128(data + '\0x01') (byte of entropy)
 	res[2], res[3] = h1, h2
 	return res
 }