… .travis.yml
… glide.yaml
… murmur_test.go
… testdata/MurmurHash3.cpp