Use hand-rolled hash function, identical to the existing algo using `m3db/stackmurmur3`

Code is all based on `twmb/murmur3` and is actually faster, without any use of `unsafe` or assembly for `amd64`
```
name                    old time/op    new time/op    delta
AddX10kX5-12              98.6ns ± 1%    67.4ns ± 0%  -31.66%  (p=0.000 n=10+9)
Contains1kX10kX5-12       89.0ns ± 1%    57.1ns ± 1%  -35.88%  (p=0.000 n=9+9)
Contains100kX10BX20-12     215ns ± 2%     173ns ± 1%  -19.35%  (p=0.000 n=10+10)

name                    old alloc/op   new alloc/op   delta
AddX10kX5-12               0.00B          0.00B          ~     (all equal)
Contains1kX10kX5-12        0.00B          0.00B          ~     (all equal)
Contains100kX10BX20-12     0.00B          0.00B          ~     (all equal)

name                           time/op
BloomFilterHash-12             29.2ns ± 1%
BloomFilterHash15-12           32.2ns ± 1%
BloomFilterHash16-12           25.5ns ± 1%
BloomFilterHashReference-12    65.9ns ± 0%
BloomFilterHashReference15-12  83.6ns ± 0%
BloomFilterHashReference16-12  67.7ns ± 0%
```